refactor: extract dataloader providers
