+++ lib/radiator/audio_meta/audio_meta.ex
@@ -12,24 +12,6 @@ defmodule Radiator.AudioMeta do
   alias Radiator.AudioMeta.Chapter
   alias Radiator.Media.ChapterImage
 
-  def data() do
-    Dataloader.Ecto.new(Repo, query: &query/2)
-  end
-
-  def query(Chapter, args) do
-    chapters_query(args)
-  end
-
-  def query(queryable, _) do
-    queryable
-  end
-
-  defp chapters_query(args) do
-    Enum.reduce(args, Chapter, fn
-      {:order, order}, query -> query |> order_by({^order, :start})
-    end)
-  end
-
   def list_chapters(%Audio{} = audio) do
     from(
       c in Chapter,
